simple fixes and refactoring (#289)
* move primitives out of runtime

* move executor out of node

* fix NodleBlock

* bump transaction_version because SignedExtra was modified

* invert tx fee split
