+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+	"executor",
 	"node",
 	"pallets/allocations",
 	"pallets/amendments",
@@ -11,6 +12,7 @@ members = [
 	"pallets/root-of-trust/rpc",
 	"pallets/root-of-trust/rpc/runtime-api",
 	"pallets/tcr",
+	"primitives",
 	"runtime",
 	"support"
 ]
