+++ executor/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "nodle-chain-executor"
+version = "2.0.0-rc6"
+authors = ["Eliott Teissonniere <git.eliott@teissonniere.org>"]
+edition = "2018"
+
+[dependencies]
+parity-scale-codec = "1.2.0"
+nodle-chain-primitives = { version = "2.0.0-rc6", path = "../primitives" }
+nodle-chain-runtime = { version = "2.0.0-rc6", path = "../runtime" }
+sc-executor = "0.8.0-rc6"
+sp-core = "2.0.0-rc6"
+sp-io = "2.0.0-rc6"
+sp-state-machine = "0.8.0-rc6"
+sp-trie = "2.0.0-rc6"
+trie-root = "0.16.0"
+frame-benchmarking = "2.0.0-rc6"
+
+[features]
+wasmtime = [
+	"sc-executor/wasmtime",
+]
+wasmi-errno = [
+	"sc-executor/wasmi-errno",
+]
+stress-test = []
